fix(ci): python-coverage-shards — split remote_repair_protected under first-party line limit
remote_repair_protected.py grew to 1505 lines (over the 1500 first-party
guardrail) after recent directive-leak/grant fixes, failing
test_first_party_code_files_stay_under_line_limit. Extract the cohesive
operator-grant and preserved-commit cluster (6 functions) into a new
remote_repair_protected_grants.py sibling module and repoint the
RunnerDelegatesMixin assignments. Behavior unchanged; the original file
drops to 1274 lines.
